requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.12
praw==5.1.0
prawcore==0.12.0
PyYAML==3.12
requests==2.18.4
SQLAlchemy==1.1.14
update-checker==0.16
urllib3==1.22
